Add more asType

+++ posthog/schema.py
@@ -919,9 +919,9 @@ class WebStatsTableQueryResponse(BaseModel):
     hogql: Optional[str] = None
     is_cached: Optional[bool] = None
     last_refresh: Optional[str] = None
-    limit: Optional[float] = None
+    limit: Optional[int] = None
     next_allowed_client_refresh: Optional[str] = None
-    offset: Optional[float] = None
+    offset: Optional[int] = None
     results: List
     samplingRate: Optional[SamplingRate] = None
     timings: Optional[List[QueryTiming]] = None
@@ -1303,9 +1303,9 @@ class QueryResponseAlternative10(BaseModel):
     hogql: Optional[str] = None
     is_cached: Optional[bool] = None
     last_refresh: Optional[str] = None
-    limit: Optional[float] = None
+    limit: Optional[int] = None
     next_allowed_client_refresh: Optional[str] = None
-    offset: Optional[float] = None
+    offset: Optional[int] = None
     results: List
     samplingRate: Optional[SamplingRate] = None
     timings: Optional[List[QueryTiming]] = None
